refactor(webapp): extract job item from table to a dedicated file
